+++ vmm/src/lib.rs
@@ -39,6 +39,7 @@ pub mod api;
 pub mod config;
 pub mod cpu;
 pub mod device_manager;
+pub mod device_tree;
 pub mod interrupt;
 pub mod memory_manager;
 pub mod migration;
